+++ CHANGELOG.md
@@ -6,6 +6,8 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 
 ## []
 ### Added
+- Choose custom pinned variants to submit to MatchMaker Exchange
+- Submit structural variant as genes to the MatchMaker Exchange
 - Added function for maintainers and admins to remove gene panels
 - Admins can restore deleted gene panels
 ### Changed